Fix gyp URL.
The v8 build itself -- at build time -- attempts to clone one of its
dependencies, gyp. (lolwut)  Because lolmips, we have to use an ancient
revision of V8, which in turn references an ancient revision of gyp as a
dependency, on code.google.com

However Google just turned off code.google.com, breaking references such as
this. Presumably Chrome has fixed this on ToT but that doesn't help us since
we can't use ToT.

So we have to patch the v8 build to replace the svn checkout with a git
clone & checkout.

Change-Id: I15f4b05a2c1a6a0872df07b0e0b55b30ef620c7a